refactor: add query filters to access request, events, access codes
