+++ frontend/src/routes/Router.tsx
@@ -7,8 +7,8 @@ import EditOrganization from '../pages/EditOrganization';
 import EditAccessCode from '../pages/EditAccessCode';
 import Login from '../pages/Login';
 import Organization from '../containers/query/OrganizationWithQueryParam';
-import RegistrationRequests from '../pages/AccessRequests';
-import AccessCodes from '../pages/AccesCodes';
+import RegistrationRequests from '../containers/query/AccessRequestsWithQueryParams';
+import AccessCodes from '../containers/query/AccessCodesWithQueryParams';
 import PrivateRoute from './PrivateRoute';
 import PublicRoute from './PublicRoute';
 import Volunteers from '../containers/query/VolunteersWithQueryParams';
@@ -19,7 +19,7 @@ import Volunteer from '../pages/Volunteer';
 import EditVolunteer from '../pages/EditVolunteer';
 import AddActivityType from '../pages/AddActivityType';
 import ActivityTypes from '../containers/query/ActivityTypesWithQueryParams';
-import Events from '../pages/Events';
+import Events from '../containers/query/EventsWithQueryParams';
 import Event from '../pages/Event';
 import AddEvent from '../pages/AddEvent';
 import EditEvent from '../pages/EditEvent';
